+++ ReleaseNotes.md
@@ -2,9 +2,9 @@
 
 ## Development version @ trunk
 
-## Development version @ branches/remove_deprecate
-
 * Remove modules deprecated in v0.10.0 (issues #1641, #1676, #1677, #1349)
+* pyhetdex.coordinates.astrometry: add option to use the IFUCen file (issue
+  #1962) and to pass dither offsets (issue #1968)
 
 ## Development version @ branches/selection_function_devel